Move Fuse to timeseries module
The Grid class is the only class that uses `Fuse` as it is used
to represent the fuse of a grid. It is not a microgrid-specific class.
Therefore, it is moved to the timeseries module.
