+++ src/frequenz/sdk/timeseries/__init__.py
@@ -37,6 +37,7 @@ Example:
 
 from .._internal._channels import ReceiverFetcher
 from ._base_types import UNIX_EPOCH, Bounds, Sample, Sample3Phase
+from ._fuse import Fuse
 from ._moving_window import MovingWindow
 from ._periodic_feature_extractor import PeriodicFeatureExtractor
 from ._quantities import (
@@ -53,6 +54,7 @@ from ._resampling import ResamplerConfig
 
 __all__ = [
     "Bounds",
+    "Fuse",
     "MovingWindow",
     "PeriodicFeatureExtractor",
     "ResamplerConfig",
